requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
tzdata==2022.6
